requirements.txt
flask
flask-login
flask-cors
flask-migrate
flask_sqlalchemy
pandas
numpy
scikit-learn
scipy
matplotlib
plotly
python-dotenv
openpyxl
gunicorn
reportlab
